+++ circuits/src/lib.rs
@@ -11,9 +11,13 @@
 #![no_std]
 
 mod encryption;
+mod recipient;
 
-/// Transaction structs, gadget, and circuit
+/// Transaction structs, and circuit
 pub mod transaction;
+
+/// Recipient Parameters
+pub use recipient::RecipientParameters;
 
 /// ElGamal asymmetric cipher
 pub use encryption::elgamal;