+++ circuits/CHANGELOG.md
@@ -7,11 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add Recipient gadget [#197]
+
 ### Changed
 
 - Remove `ViewKey` from `TxOutputNote::new()` parameters [#191]
 - Make `rng` the first param in `TxInputNote::new` [#189]
 - Rename `crossover` to `deposit` [#190]
+- Remove recomputation of `value_commitment` in `TxOutputNote::New()`
+- Rename `skeleton_hash` to `payload_hash` [#188]
+- Make `TxCircuit` to use the Recipient gadget
 
 ## [0.1.0] - 2024-05-22
 
@@ -29,6 +36,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update `poseidon-merkle` to v0.6 [#179]
 
 <!-- ISSUES -->
+[#197]: https://github.com/dusk-network/phoenix/issues/197
+[#188]: https://github.com/dusk-network/phoenix/issues/188
 [#191]: https://github.com/dusk-network/phoenix/issues/191
 [#190]: https://github.com/dusk-network/phoenix/issues/190
 [#189]: https://github.com/dusk-network/phoenix/issues/189